refactor(rust): move `auth0::node` to `util`
