+++ implementations/rust/ockam/ockam_command/src/util/mod.rs
@@ -23,6 +23,7 @@ use crate::{CommandGlobalOpts, OutputFormat};
 
 pub mod api;
 pub mod exitcode;
+pub mod node;
 pub mod startup;
 
 mod addon;
